container/python_launcher/requirements.txt
azure-servicebus==7.8.3
azure-identity==1.12.0

requests==2.28.2